requirements.txt
streamlit==1.39.0
python-pptx==0.6.23
openai==0.28.1
python-dotenv==1.0.1